backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
